Catch errors when attempting to initialize JIRAClient.
If a bug tracker is configured incorrectly, JIRAClient won't be able to
initialize, and will unhelpfully throw a JSONDecodeError instead of
something more useful. The only place where this manifests is in the bug
infobox (since we otherwise are just creating a link to the bug). The
result of this was that crash e-mails were being sent to the admin
whenever the infobox was triggered (with no visible effect to the user).

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/10144/