Fix tracebacks occurring in GetOriginalFileTests
There were tracebacks occurring in the `GetOriginalFileTests` due to the
`user` not being present on the `django.http.HttpRequest` that was used
in logging, causing an `AttributeError`. We now provide an anonymous
user on the request.

Testing Done:
Ran unit tests and did not see any tracebacks.

Reviewed at https://reviews.reviewboard.org/r/10175/