Correctly serialize review requests with only-fields
If the `dvcs_feature` is disabled and a request is made to the review
request WebAPI resource with the `only-fields/only_fields` query/post
parameter set to a subset of fields that does not include the
`created_with_history` field, the server would raise an exception and
return a HTTP 500. This was because the field was already excluded from
the set of serialized fields.

We now are more defensive about removing this field so as not to raise a
`KeyError` if the field is not in the dictionary.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/10169/