Update hosting service paginators for type safety and modern docs.
This updates the hosting service paginator classes to check that the
subclasses are providing the correct values when returning a page of
results. This gives us assurances that the values will be compatible
across Python 2.x and 3.x.

The documentation has been updated to explicitly state the valid types.
The rest of the documentation for the file has also been modernized to
help make usage of these much more clear.

Testing Done:
Unit tests pass.

Built the docs and checked them for obvious errors.

Reviewed at https://reviews.reviewboard.org/r/10203/