Add docs for SELinux boolean httpd_can_connect_ldap
In a recent thread on the mailing list, a user noted that SELinux was denying
access to LDAP and that this could be alleviated by setting another SELinux
boolean. As this is useful information, we should include it in the official
documentation.

Testing Done:
None in particular; changes are purely in text.

Reviewed at https://reviews.reviewboard.org/r/10201/