Move commit validation utilities into commit_utils
The utilities for serialization, deserialization, and generation of
commit validation metadata have been moved into the `commit_utils`
module. Previously, the (de-)serialization logic was implemented in a
multitude of places and the generation was intrinsically tied to the
`ValidateDiffCommitResource`.

However, we will soon require validation metadata in the
`DraftDiffResource` for finalization of commit series, and so to not
reimplement this all in another place, it has been refactored into the
`diffviewer.commit_utils` module.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/10227/