Display the selected revision's commits in the diffviewer
The diffviewer now displays the list of commit messages and authors and
they are updated when the selected revision changes. Currently the
commit list doesn't support interdiffs and an error message is
displayed instead. This will be addressed in a future patch.

Testing Done:
- Ran unit tests.
- Ran JS tests.

Reviewed at https://reviews.reviewboard.org/r/10125/