Fix text box failing to Linkify paragraph children
When making a paragraph with multiple newline links in a text box and hitting
'ok' only the first link would be linkified.

This was tested and it was found that the paragraph children were not handling
siblings correctly such that if one is replaced it causes the loop variable
'node' becomes stale when referencing next siblings.

The best method to get around this was to use the fact that all the child nodes
are available to be iterated in order. By doing so we avoid dealing with
pointer references which when replaced were causing the issues.

Co-authored-by: Stuart Caie <stuart.caie@oracle.com>

Testing Done:
Ran JS tests.
Ran unit tests.

Tested by inputting paragraphs of links in different amounts.
Tested by putting plain text above and below paragraphs of links inclusive
and exclusively with multiple links and singular links.

Bugs closed: 4685

Reviewed at https://reviews.reviewboard.org/r/10152/