Fix Bitbucket repo name description error.
When adding a repository as an admin, and choosing Bitbucket as a hosting
service, an incorrect repository name description displays. In the "Repository
Name" description, when using the "Personal" repository plan, the description
would read "The username of the user who owns the repository. This is the
<repo_name> in https://bitbucket.org/<username>/<repo_name>/". This isn't
correct.

It now reads "The name of the repository. This is the <repo_name> in
https://bitbucket.org/<username>/<repo_name>/".

Testing Done:
Tested by refreshing the page, and confirmed that the "Repository Name"
description is the new description. Also, navigated to
http://localhost:8080/admin/, clicked on "Add +" under Repositories, and
confirmed that on http://localhost:8080/admin/db/scmtools/repository/add/ the
description text has changed.

No other description fields have changed.

Bugs closed: 4664

Reviewed at https://reviews.reviewboard.org/r/10150/