Serialize primary keys in DiffCommits and history entries
When doing an interdiff, it is possible that there could be two commits
in the `RB.DiffCommitCollection` that have the same `commitId` property.
However, they may have different properties (however unlikely), so we
now serialize commits with their primary keys and use them for lookup in
history entries instead of the commit ID. This was especially prevalent
in test cases where the commit IDs were not actually representative
(i.e., a hash) of the content of the commit.

Testing Done:
- Ran unit tests.
- Ran JS tests.

Reviewed at https://reviews.reviewboard.org/r/10128/