Generate diffs between base and tip commits
The `get_diff_files` utility can now generate a list of files to diff
that exist between two commits in a diffset. This allows us to generate
diffs between two commits (and therefore to see the diff of any
individual commit). There is currently no UI support for this behaviour,
but that is comming in a future patch.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/10130/