Sort the serialized dictionary of avatar URLs.
The `{% avatar_urls %}` template tag looks up the avatar URLs for a user
and returns a JSON payload that represents them for use in a template.
Depending on the version of Python being used, the order of keys would
be different, making it annoying to unit test.

This change simply sorts the keys when dumping to JSON, giving us a
predictable order.

Testing Done:
Unit tests pass on Python 2.7, 3.5, and 3.6.

Reviewed at https://reviews.reviewboard.org/r/10259/