Update hosting services to use HostingServiceHTTPResponse.
This is a small update to the hosting service implementations that
switch from the `(data, headers)` form of HTTP responses to
`HostingServiceHTTPResponse`. It's a small step toward modernizing the
hosting services (which would need to come as a series of larger
changes), but gets us closer to a simpler codebase.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10217/