Include branch field in review request emails
With a lot of review request emails, prioritizing review requests emails
can sometimes be difficult. One request we've had is to include the branch name
in the email, so people can quickly get a sense of where a change is going to
land. This change implements that.

Testing Done:
Manually checked generated review request emails for the inclusion of the branch
field.

Bugs closed: 4707

Reviewed at https://reviews.reviewboard.org/r/10147/