Fix the preview password changed e-mail view
The `preview_password_changed_mail` view was (a) not correctly
referenced in the url patterns and (b) was not providing the list of TO
addresses as, well, a list. Both of these issues have been resolved and
now the view renders properly.

Testing Done:
- Ran unit tests.
- Viewed the password preview view.

Reviewed at https://reviews.reviewboard.org/r/10274/