Serialize commit count in diff resource
We now include the `commit_count` in the payload for the diff resource
when the DVCS feature is enabled. This allows RBTools to check for
review requests created with history support when doing `rbt patch`
without making an additional request to either the commit list resource
or the review request resource when unnecessary.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/10282/