Allow repository names to be 255 characters long
When creating a repository, the name field previously allowed the name to be 65
characters long. Once the name got to 65 characters, you would not be able to
type any additional characters.

Now, you are allowed to specify a name that is less than or equal to 255
characters long.

Testing Done:
- Ran `./reviewboard/manage.py evolve --execute evolve` to update the database.
- Restarted my server
- Manually Tested by
      - Clicked on the add button for Repositories
      - Specified a name that was 255 characters long
      - Clicked save
      - Got no Errors
      - Verified that the name shown in my list of repositories was
        255 characters long
- Added a unit test to verify creating a repository name with 255 characters
  works, and does not produce errors

Bugs closed: 4730

Reviewed at https://reviews.reviewboard.org/r/10153/