Allow publishing review requests without updating diff
The `CommitListField` was assuming the field changed even when a new
diff wasn't uploaded, resulting it it failing to record change entry
information and, ultimately, prevent publishing.

We now only mark the field as changed when a new diff is present.

Testing Done:
- Published a review request created with history when the diff was
  not updated.
- Published a review request created with history when the diff was
  updated.

Reviewed at https://reviews.reviewboard.org/r/10283/