Properly render review request details page when draft has no commits
When a draft review request was added without an associated diff, we
would fail to render the reviewable page model data due to assuming that
the review request details (which might be a draft or a review request)
would walways have a diffset attached. This assumption has been
corrected.

Testing Done:
Am now able to view a review request details page for a review request
created without commit history support when the current draft does not
have a diff attached.

When there is a draft with a diffset attached, the draft commits are
shown instead.

Reviewed at https://reviews.reviewboard.org/r/10284/