Fix a review request page caching issue when setting close descriptions.
Attempting to modify a review request's close description and then
reloading the page gave the impression that the message didn't save.
This was because the timestamp for the accompanying ChangeDescription
wasn't being factored into the ETag for the page. A shift-reload would
fix this, as would any change affecting other pieces of the ETag.

We now include the timestamp for the ChangeDescription as part of the
ETag, ensuring that any reloads will show the new content immediately.

Testing Done:
Unit tests pass.

Manually tested that old content was no longer being cached on reload.

Reviewed at https://reviews.reviewboard.org/r/10349/