Archive review requests from inbox icon
Made the inbox icon on review request pages a shortcut for
archiving/unarchiving the review request. The idea behind this change is to
make the archiving/unarchiving of requests a little bit faster for advanced
users.

The shortcut was implemented by adding archive/unarchive click handlers to the
icon. A title attribute was also added to the icon which dynamically changes to
"Archive review request" or "Unarchive review request". This should help with
accessibility and general ease of use.

Testing Done:
There were no changes made to any Python files so the Python tests were not ran
for this change.

The JS tests were ran for these changes. These changes were implemented off the
master branch version of ReviewBoard which had 6 failures for the JS tests.

The JS tests when ran on this update had the same 6 failures as in the master
branch JS tests.

Reviewed at https://reviews.reviewboard.org/r/10154/