Archive repositories when sending a DELETE to the repository resource.
We've had support for hiding repositories through the API by issuing
HTTP DELETE requests to the API resource since the 1.6.x timeframe.
After that, archiving of repositories was introduced, which the API
gained support for through a PUT with `archive_name=1`, which worked but
probably should have been done through a DELETE request.

Now, when performing a DELETE, we archive the repository instead. This
has the same general effect as what we had before, with the benefits
that the repository name could be reused afterward.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10374/